requirements.txt
spotipy
python-dotenv
google-api-python-client
rapidfuzz
ytmusicapi
google-auth-oauthlib
google-api-python-client